requirements.txt
boto3==1.7.41
botocore==1.10.41
certifi==2018.4.16
chardet==3.0.4
click==6.7
coverage==4.5.1
docutils==0.14
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
nose2==0.7.4
Pillow==5.1.0
python-dateutil==2.7.3
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
